Error register and no more PROGMEM variables
added an error "register" (bitwise errorflags in one integer)
deleted all PROGMEM variables

+++ Mechaduino/State.h
@@ -23,7 +23,7 @@ extern volatile bool frequency_test;       // flag for frequency test
 extern volatile bool streaming;	  		// flag for data streaming
 
 
-extern volatile byte error_register;
+extern volatile int16_t error_register;
 
 //---- PID Gains ----
 extern volatile int int_Kp;
